Ugly workaround for latex -outdir
As mentioned by https://github.com/kjellmf/dot2tex/issues/52, dot2tex fails when the tex document is compiled with a -outdir CLI switch because it is not informed that the .dot file was output in a different place, as dot2texi \writes will implicitly go there. This ugly hack allows to load dot2texi with a [inputdir=./build/] option that should be set to match the latex -outdir.